Front end Styling for Generate Bio
